Add loop for generating photo classes
This commit adds a new file "11-loop_photos.scss" that imports the "11-photos.scss" file and creates a loop to generate photo classes based on the names in the $list-names variable. Each photo class is assigned a background image using the corresponding photo name. This feature enhances the flexibility and scalability of the photo styling in the project.