Refactor flexbox styles for hero section and latest news
This commit refactors the flexbox styles for the hero section and latest news. It adjusts the padding in the section-hero class to improve the layout and adds a flex-direction: row-reverse property to the section-latest-news class to change the order of the news items. Additionally, it introduces a new section-services class with a flex-wrap: wrap property to enable wrapping of service items. These changes enhance the overall design and responsiveness of the website.