Refactor flexbox styles for navbar and sections
This commit refactors the flexbox styles for the navbar, hero section, and latest news. It removes unnecessary margin properties from the .nav .nav-item class and adjusts the font styles for better readability. Additionally, it introduces a new section-services class with a flex-wrap: wrap property to enable wrapping of service items. These changes enhance the overall design and responsiveness of the website.